Fixes #70 - CSV and markdown files are saved to target directory
